Refactor game scripts into modular structure
