+++ web/index.html
@@ -26,6 +26,6 @@
         </div>
       </div>
     </div>
-    <script src="./js/game.js"></script>
+    <script type="module" src="./js/game.js"></script>
   </body>
 </html>
