+++ web/js/coop.js
@@ -1,0 +1,5 @@
+import { GameWorld } from './single-player.js';
+
+export function createCoopWorld(options) {
+  return new GameWorld({ ...options, mode: 'coop' });
+}
